+++ common/src/main/java/com/gitlab/cdagaming/craftpresence/core/config/category/Advanced.java
@@ -56,7 +56,6 @@ public class Advanced extends Module implements Serializable {
     public boolean allowDuplicatePackets = false;
     public int maxConnectionAttempts = 10;
     public boolean enableClassGraph = false;
-    public boolean useClassLoader = Constants.USE_CLASS_LOADER;
 
     public Advanced(final Advanced other) {
         transferFrom(other);
@@ -97,7 +96,6 @@ public class Advanced extends Module implements Serializable {
             allowDuplicatePackets = data.allowDuplicatePackets;
             maxConnectionAttempts = data.maxConnectionAttempts;
             enableClassGraph = data.enableClassGraph;
-            useClassLoader = data.useClassLoader;
         }
     }
 
@@ -121,7 +119,6 @@ public class Advanced extends Module implements Serializable {
             case "allowDuplicatePackets" -> allowDuplicatePackets;
             case "maxConnectionAttempts" -> maxConnectionAttempts;
             case "enableClassGraph" -> enableClassGraph;
-            case "useClassLoader" -> useClassLoader;
             default -> null;
         };
     }
@@ -181,9 +178,6 @@ public class Advanced extends Module implements Serializable {
                 case "enableClassGraph":
                     enableClassGraph = (Boolean) value;
                     break;
-                case "useClassLoader":
-                    useClassLoader = (Boolean) value;
-                    break;
                 default:
                     break;
             }
@@ -218,8 +212,7 @@ public class Advanced extends Module implements Serializable {
                 Objects.equals(other.playerSkinEndpoint, playerSkinEndpoint) &&
                 Objects.equals(other.allowDuplicatePackets, allowDuplicatePackets) &&
                 Objects.equals(other.maxConnectionAttempts, maxConnectionAttempts) &&
-                Objects.equals(other.enableClassGraph, enableClassGraph) &&
-                Objects.equals(other.useClassLoader, useClassLoader);
+                Objects.equals(other.enableClassGraph, enableClassGraph);
     }
 
     @Override
@@ -232,7 +225,7 @@ public class Advanced extends Module implements Serializable {
                 allowEndpointIcons,
                 serverIconEndpoint, playerSkinEndpoint,
                 allowDuplicatePackets, maxConnectionAttempts,
-                enableClassGraph, useClassLoader
+                enableClassGraph
         );
     }
 }